requirements.txt
openai>=1.0.0
hyperliquid-python-sdk>=0.1.0
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
requests>=2.31.0
loguru>=0.7.0
pydantic>=2.0.0
tenacity>=8.2.0
flask>=3.0.0
psycopg2-binary>=2.9.0
